BE-61: change userId int to long

+++ src/main/java/az/gdg/msarticle/client/MsAuthClient.java
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.PutMapping;
 @FeignClient(value = "ms-auth-client", url = "${service.url.ms-auth}")
 public interface MsAuthClient {
     @PutMapping("/user/popularity/{userId}")
-    void addPopularity(@PathVariable("userId") Integer userId);
+    void addPopularity(@PathVariable("userId") Long userId);
 }
